Support Selenium 4
From Selenium 4, `chromedrive` path is specified in `executable_path`
of `webdriver.service.Service` before passing it to `service` param
of `webdriver.Chrome`. `chrome_options` is renamed to `options` as
well.

Tested with 3.14.0 and latest version 4.10.0

Please noted: there is the issue with 3.14.1 and 3.141.0 but not
be counted in this PR.